wrote working createAdmin

+++ backend/routes/users.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { getAllUsers, createUser, deleteAll, getUserByUsername, updatePassword } = require('../controllers/users');
+const { getAllUsers, createUser, createAdmin, deleteAll, getUserByUsername, updatePassword } = require('../controllers/users');
 const { login } = require('../controllers/login');
 const { logout } = require('../controllers/logout');
 
@@ -14,6 +14,10 @@ router.get('/', (req,res) => {
 
 router.post('/register', (req,res) => {
     createUser(req, res);
+});
+
+router.post('/registerAdmin', (req,res) => {
+    createAdmin(req, res);
 });
 
 // add forwardAuth
